requirements.txt
faiss-cpu
sentence-transformers
groq
pandas
openpyxl
numpy
